Removing docker input, using deg F for IPMI.
Try as I might I couldn't find a way to get ipmitool to use Fahrenheit
instead of Celsius. I looked in the ipmitool source (which has "degrees
F" but no option to force it), in the ipmi0 kernel module source, and in
my BIOS settings. Nothing.

The solution is to write a wrapper AWK program that does the conversion
for Telegraf.

Won't work until (hopefully)
https://github.com/influxdata/telegraf/pull/2498 is merged.

Not collecting Docker metrics since it requires adding Telegraf to the
"docker" group to give it access to the Docker unix socket. I didn't
really like that from a security perspective. Better to just give up on
that.